Refactor: Moves MdDialogRender to MdDialog folder
