+++ src/components/MdDialog/MdDialogRender.js
